Updates to build static/modular
