+++ Makefile
@@ -1,17 +1,50 @@
 # Makefile - just builds the binary, for dev mainly
 
-.PHONY: clean test generate testbot
+.PHONY: clean test generate testbot modular static
 
 commit := $(shell git rev-parse --short HEAD)
 
+# MODULES = connectors/slack.so connectors/rocket.so connectors/terminal.so \
+#	goplugins/duo.so goplugins/knock.so goplugins/totp.so brains/dynamodb.so
+
+MODULES = goplugins/knock.so
+
 GOOS ?= linux
-CGO ?= 1
+CGO ?= 0
+
+modular: CGO = 1
+modular: BUILDTAG = modular
+modular: gopherbot $(MODULES)
+
+static: gopherbot
 
 gopherbot: main.go bot/* brains/*/* connectors/*/* goplugins/*/* history/*/*
-	CGO_ENABLED=${CGO} GOOS=${GOOS} GOARCH=amd64 go build -mod vendor -ldflags "-X main.Commit=$(commit)" -tags 'netgo osusergo static_build' -o gopherbot
+	CGO_ENABLED=${CGO} GOOS=${GOOS} GOARCH=amd64 go build -mod vendor -ldflags "-X main.Commit=$(commit)" -tags "netgo osusergo static_build $(BUILDTAG)" -o gopherbot
+
+# modules
+connectors/slack.so: connectors/slack-mod.go connectors/slack/*.go
+	GOOS=${GOOS} GOARCH=amd64 go build -mod vendor -o $@ -buildmode=plugin -tags 'netgo osusergo static_build modular' $<
+
+connectors/rocket.so: connectors/slack-mod.go connectors/rocket/*.go
+	GOOS=${GOOS} GOARCH=amd64 go build -mod vendor -o $@ -buildmode=plugin -tags 'netgo osusergo static_build modular' $<
+
+connectors/terminal.so: connectors/slack-mod.go connectors/terminal/*.go
+	GOOS=${GOOS} GOARCH=amd64 go build -mod vendor -o $@ -buildmode=plugin -tags 'netgo osusergo static_build modular' $<
+
+goplugins/duo.so: goplugins/duo-mod.go goplugins/duo/*.go
+	GOOS=${GOOS} GOARCH=amd64 go build -mod vendor -o $@ -buildmode=plugin -tags 'netgo osusergo static_build modular' $<
+
+goplugins/knock.so: goplugins/knock-mod.go goplugins/knock/*.go
+	GOOS=${GOOS} GOARCH=amd64 go build -mod vendor -o $@ -buildmode=plugin -tags 'netgo osusergo static_build modular' $<
+
+goplugins/totp.so: goplugins/totp-mod.go goplugins/totp/*.go
+	GOOS=${GOOS} GOARCH=amd64 go build -mod vendor -o $@ -buildmode=plugin -tags 'netgo osusergo static_build modular' $<
+
+brains/dynamodb.so: brains/dynamodb-mod.go brains/dynamodb/*.go
+	GOOS=${GOOS} GOARCH=amd64 go build -mod vendor -o $@ -buildmode=plugin -tags 'netgo osusergo static_build modular' $<
 
 clean:
-	rm -f gopherbot
+	rm -f gopherbot $(MODULES)
 
 # Run test suite
 test:
@@ -21,5 +54,6 @@ test:
 generate:
 	go generate -v --tags 'test integration netgo osusergo static_build' -mod vendor ./bot/
 
+# Terminal robot that emits events gathered, for developing integration tests
 testbot:
 	CGO_ENABLED=0 GOOS=${GOOS} GOARCH=amd64 go build -mod vendor -tags 'netgo osusergo static_build test' -o gopherbot
