+++ goplugins/knock/static.go
@@ -1,0 +1,9 @@
+// +build !modular
+
+package knock
+
+import "github.com/lnxjedi/gopherbot/bot"
+
+func init() {
+	bot.RegisterPlugin("knock", knockhandler)
+}
