+++ main_static.go
@@ -1,0 +1,8 @@
+// +build !modular
+
+package main
+
+// blank imports used in static builds
+import (
+	_ "github.com/lnxjedi/gopherbot/goplugins/knock"
+)
